Skip structurally incompatbile overload candidates
Calls to overloaded functions result in a pair of constraints. One
associates the callee's type with a synthesized function type, the
other further constrain the callee's type to an overload set.

The purpose of this change is to leverage the structural information
provided by the former constraint to narrow the set of choices when
the second is being solved.