Fix Windows CI again (#246)
Fixes #243 to the extent possible in this project.

Apparently the problem has to do with the --enable-code-coverage option, which we're not currently
taking advantage of on Windows.  This change turns it only for Linux platforms, where the coverage action is designed to run.  Also does a little bit more to make sense of our text matrix.  GH actions are a mystery!