Generate CXX entry point
If the Val program contains a `main` function, we generate a CXX `main`
in the global namespace that would call the `main` from the proper module.
If the Val `main` returns an `Int`, that value is returned from the
program. Otherwise, we discard the result of the `main` function and
return 0.

With this change, we can compile the CXX code for the factorial example.