Simplify the property map storing generic environments
Generic environments were stored in a property map keeping track
of in-progress requests to catch circular dependencies. However,
there was no recovery mechanism in place to handle such a failure.

It is probably better to handle circular dependencies occurring
during the construction of generic environments using the
declaration request status of the associated declaration.