Remove the 'deinit' instruction
At the IR level, deinitialization should be expressed by an explicit
call to a sink method rather than a custom instruction and implicit
destruction after last use should be lowered as a call to the method
implementing `Destructible.destory`. Before the support for this
trait is implemented, we can replace uses of `deinit` by a `load`.