Debug CI on Ubuntu
For some reason, CI is not able to run anythin in the postCreateCommand
phase of the dev container setup. That makes it impossible to properly
configure pkgconfig so that swiftpm can find LLVM headers.

To fix this problem, this patch moves the configuration of pkgconfig in
the the CI job description, running it right before 'swift build'.
Unfortunately, that means one has to manually run 'make-pkgconfig.sh' if
they want to use a dev container locally.