Use a different action to setup swift on Linux. (#908)
An attempt to stomp out the spurious format-check failures we've been seeing, that [die in setup-swift](https://github.com/swift-actions/setup-swift/issues/591).

An alternative would be to run the format check on MacOS, but that's expected to be slower.  This analysis <https://github.com/swift-actions/setup-swift/issues/591#issuecomment-1684602660> explains why I think it would work.

Another alternative would be to run it on Linux in the devcontainer, where Swift is already set up.