Pass null pointers as arguments of zero-sized parameters
The purpose of this change is to remove some of the noise caused
by pointless allocations of `Void` passed as return values. In
general, it should be possible to avoid "allocating" storage for
any zero-sized type, as Hylo makes any guarantee on the values of
parameter addresses.

Note that because null pointers may be passed as argument, we can
no longer add the `nonnull` and `noundef` attributes.