Use the type of the whole expression to lower upcasts
As mentioned in #1158, the type checker currently doesn't reify
the type of expressions appearing in a type hint position. That
means we can't reliably expect `program[e].cast` to be a closed
type. However, we can use the type of the whole expression since
inference always concludes that it must be equal to the (reified)
type of cast's RHS.