Reuse type inference to evaluate type expressions
In early versions of the AST, type and value expressions had
distinct ASTs, which made it easier to compute the "value" of a
type annotation. The distinction got erased but the code processing
expressions occurring as type annotations was not merged with the
code inferring the type of value expressions, causing to code
duplication. This commit taclkes this problem by reusing the logic
of type inference to process type annotations.