Fix the specialization of associated types
When an associated type rooted at a trait receiver is referred to as a
member of a trait `P`, the root of that type must be substituted by the
receiver of `P`, regardless of the trait declaring the associated type.
The goal is to allow monomorphization to properly substitute associated
types without requiring a substitution defined for every trait in the
refinement cluster of `P`. For example:

    trait P {
      type A
      fun f() -> A
    }
    trait Q: P {
      fun g() -> A { self.make_a() }
    }

Here, a reference to `A` in the context of `Q` must be rooted at the
`Q`'s receiver rather than `P`'s.