Add tests for concurrency code.
The concurrency code need to link against multiple libraries that are
not automatically build in the Hylo project. For this reason, we compile
these tests only until the "emit LLVM" step, without trying to generate
the associated binary. With this, we can check if there are any errors
compiling the Hylo code, but we still can't check if the execution is
correct.