Modify the layout of lambdas
The layout is modified so that the structural properties of a lambda
don't depend on the size of its environment. Now, a lamibda is always
represented as a pair. The first element is a function pointer and the
second is a record representing its environment and whose parts are no
longer "exploded" directly in the layout of the lambda.

As a result, it is always defined to access the address of a lambdas'
environment, even if the lambda's thin.