Fix unification of union types
The algorithm is a little naive. We unify each element of the LHS with
each element of the RHS and record if at least one pair succeeded. The
intuition is to try all different ways in which the LHS can be uniified
with the RHS "at once", without having to backtrack.

This strategy will fail to unify unions whose elements have identical
constructors but different variables. For example, it will fail for
`Union<A<%1>, A<%2>>` and `Union<A<T>, A<U>>` because both variables
will be assumed to be either `T` or `U` and the algorithm won't "try"
to discover other substitutions.