Modify the typing rules for mutating method bundles
The return type of method bundles with mutating variants used to
be a tuple `{self: A, T}`, with `A` the receiver and `T` the return
value of the `inout` or `set` variant. This approach is verbose for
the user and creates complications in the type checker. This patch
modifies that rule so that the return type is expressed in terms of
what an `inout` or `set` variant returns.

Type checking further requires to determine the result of a call to
a method bundle before dataflow analysis. Before this patch, the
checker would always pick the declared output type of the method,
which was wrong for in-place calls. Now it uses mutation markers to
determine *how* a method is called and picks the right return type.