Add private linkage to Hylo's entry point
The "main" function of the program in LLVM IR is a function calling
Hylo's entry point (i.e., the `public main fun` in Hylo sources).
With optimizations enabled, the latter is typically inlined in the
former, leading to unnecessary code duplication. Since it is not
possible to call Hylo's entry point from a library, through, we can
add private linkage to that function.