Redesign 'union_switch' to avoid raw accesses
The previous design of the instruction did not require the operand
to be the result of a 'let' access, which would cause problems
during access reification. The new design consumes a discriminator
value that must have been obtained from 'union_discriminator'.

An alternative design could be to require that the first operand be
an access that would be passed as a basic block arguments to the
different targets. This approach makes it more difficult to track
lifetime dependencies, though.