Move `uses` from `Module` to `Function`.
The uses of an operand shall always be part of the function containing
the operand. Moving the `uses` cache from `Module` to `Function`
improves locality of this cache and allows changing the IR of a function
more easily without necessarily needing to mutate the top-level `Module`
object.

Changes:
- Move `uses` cache from `Module` to `Function`
- Change all code to use the corresponding `Function` object instead of
the `Module` object.