Rename `InstructionID` to `AbsoluteInstructionID`
Most of the time we are interested in instructions, we care about them
in the context of a function. Ideally, we should not refer to
instructions outside of functions, or across functions. But, using a
`Function.ID` as part of `InstructionID` will make all the functions
that use `InstructionID` to not be used on modules, and not on `Function`
objects.

Rename `InstructionID` to `AbsoluteInstructionID`. We want to keep the
short and good name for referencing instructions in the scope of a
function object.