Make Depolymorphization not need the module anymore.
We split the processing in depolymorhization in three phases:
1. observation -- check all the types and function references we have
in a function
2. monomorphization -- transform the types / function references (using
information from the modile)
3. instruction transformation -- just transforms instructions making
sure that previous made monomorphization is applied

To support step 1, we added `InstructionObserver` that corresponds to
`InstructionTransfomers`, but it only observes the parts of instructions
in a function, without making any transformation.

The goal is to ensure that instruction transformation doesn't need to
perform meonomorphization in its processing, as this would depend on
Module.